+++ packages/component-library/src/components/feedback-indicator/mt-snackbar/mt-snackbar.stories.ts
@@ -11,7 +11,7 @@ import type { SlottedMeta } from "@/_internal/story-helper";
 export type MtSnackbarMeta = SlottedMeta<typeof MtSnackbar, "">;
 
 const meta: Meta<typeof MtSnackbar> = {
-  title: "Components/Feedback Indicator/mt-snackbar",
+  title: "Components/mt-snackbar",
   component: MtSnackbar,
 };