+++ packages/component-library/src/components/form/mt-select/mt-select.interactive.stories.ts
@@ -7,7 +7,8 @@ import meta, { type MtSelectMeta, type MtSelectStory } from "./mt-select.stories
 
 export default {
   ...meta,
-  title: "Interaction Tests/Form/mt-select",
+  title: "Components/mt-select/Interaction tests",
+  tags: ["!autodocs"],
 } as MtSelectMeta;
 
 export const TestRemovalOfSelectionByList: MtSelectStory = {