+++ packages/component-library/src/components/layout/mt-empty-state/mt-empty-state.interactive.stories.ts
@@ -6,7 +6,8 @@ import MtButton from "@/components/form/mt-button/mt-button.vue";
 
 export default {
   ...meta,
-  title: "Interaction Tests/Layout/mt-empty-state",
+  title: "Components/mt-empty-state/Interaction tests",
+  tags: ["!autodocs"],
   component: MtEmptyState,
 } as Meta;